Internship report
The Bachelor of Statistics Study Program implements the PKL program as stated in the curriculum. In producing competent and qualified statistics practitioners who are able to manage data, starting from collecting, processing, presenting data according to the needs of the world of work. Each student can choose a PKL location from a government or private agency that carries out data processing. One of the government agencies that can be used as a place for street vendors' activities is the Pesisir Selatan Regency Communication and Informatics Service (Kominfo).
The author chose the Department of Communication, Informatics, Pesisir Selatan Regency because it is close to where he lives so that he can carry out street vendors optimally.